# docs/api/virtualpet-openapi.yaml
openapi: 3.1.0
info:
  title: VirtualPet API
  version: 1.0.0
  description: >
    E-commerce de productos para mascotas. Cubre dos aplicaciones cliente:
    Marketplace (clientes finales) y Backoffice (operarios de depósito).


    Convenciones transversales:

    - **Dinero:** importes como string decimal con escala fija (BigDecimal en backend);
      moneda en ISO 4217 (`java.util.Currency`). Se serializan como string para no perder
      precisión en clientes JavaScript.

    - **Paginación:** por cursor opaco (`cursor`, `limit`) en los listados.

    - **Errores:** RFC 7807 (`application/problem+json`).

    - **Idempotencia:** header `Idempotency-Key` obligatorio en operaciones de pago.

    - **Correlación:** header `X-Request-Id` en toda request/response.

    - **Seguridad:** rate limiting (Redis) en endpoints de auth; `429` con `Retry-After`.
servers:
  - url: https://api.virtualpet.com/api/v1
    description: Producción
  - url: http://localhost:8080/api/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      description: >
        Autentica con email y contraseña y devuelve los tokens. Si la petición incluye la cookie
        CART_SESSION, el carrito anónimo se fusiona con el carrito del usuario (sumando cantidades
        de SKUs duplicados), se elimina el carrito anónimo y la cookie se borra (Set-Cookie con
        Max-Age=0). No crea carrito de usuario salvo que haya items que fusionar (creación lazy).
      security: []
      parameters:
        - { $ref: '#/components/parameters/CartSession' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Autenticado
          headers:
            Set-Cookie:
              schema: { type: string }
              description: >
                Presente solo si llegó la cookie CART_SESSION; la borra tras la fusión
                (CART_SESSION=; Max-Age=0).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout
      description: Invalida el refreshToken (blacklist en Redis). El accessToken caduca solo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '204': { description: Sesión cerrada }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refrescar access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Nuevo access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  tokenType: { type: string, examples: [Bearer] }
                  expiresIn: { type: integer, examples: [900] }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/register/customer:
    post:
      tags: [auth]
      summary: Registro de cliente
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                firstName: { type: string }
                lastName: { type: string }
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/register/employee:
    post:
      tags: [auth]
      summary: Registro de operario (ADMIN)
      description: Alta de operario de backoffice. Solo ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                firstName: { type: string }
                lastName: { type: string }
      responses:
        '201':
          description: Operario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/me:
    get:
      tags: [auth]
      summary: Datos del usuario autenticado
      responses:
        '200':
          description: Perfil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [auth]
      summary: Actualizar datos personales
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/password/change:
    post:
      tags: [auth]
      summary: Cambiar contraseña (autenticado)
      description: Exige la contraseña actual; invalida el resto de las sesiones.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password }
      responses:
        '204': { description: Contraseña cambiada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /cart:
    get:
      tags: [cart]
      summary: Obtener el carrito
      description: >
        Devuelve el carrito. Si la petición está autenticada (Bearer token) se resuelve por usuario;
        en caso contrario se identifica por la cookie CART_SESSION (que lleva solo el id de sesión;
        el contenido vive en Redis). Sin usuario ni cookie responde un carrito vacío. Creación lazy:
        responde 200 aunque el carrito no exista.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - { $ref: '#/components/parameters/CartSession' }
      responses:
        '200':
          description: Carrito (vacío o con items)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
  /cart/items/{skuId}:
    put:
      tags: [cart]
      summary: Fijar cantidad de un SKU (upsert)
      description: >
        Fija la cantidad exacta. Dispara la creación lazy del carrito si no existe. Idempotente.
        Si la petición es anónima y no trae la cookie CART_SESSION, se genera una sesión nueva y se
        devuelve en Set-Cookie (solo el id de sesión; el contenido vive en Redis); las peticiones
        siguientes la reenvían automáticamente.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - { $ref: '#/components/parameters/SkuId' }
        - { $ref: '#/components/parameters/CartSession' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Item resultante
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetCartSession' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartItemQuantity' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [cart]
      summary: Eliminar un SKU del carrito
      description: >
        Elimina el producto por completo. Idempotente (éxito aunque no estuviera). Las peticiones
        anónimas se identifican por la cookie CART_SESSION (solo el id de sesión; el contenido vive
        en Redis).
      security:
        - bearerAuth: []
        - {}
      parameters:
        - { $ref: '#/components/parameters/SkuId' }
        - { $ref: '#/components/parameters/CartSession' }
      responses:
        '204': { description: Eliminado (o ya no estaba) }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SkuId:
      name: skuId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CartSession:
      name: CART_SESSION
      in: cookie
      required: false
      schema: { type: string }
      description: >
        Id de sesión del carrito anónimo. La cookie lleva únicamente el id; el contenido del
        carrito vive en Redis (cart:anon:{id}), nunca en la cookie. Se emite en la primera
        mutación anónima y el navegador la reenvía automáticamente.
  headers:
...
    SetCartSession:
      schema: { type: string }
      description: >
        Cookie CART_SESSION (HttpOnly) con el id de sesión del carrito invitado — solo el id,
        el contenido vive en Redis. Presente solo en la primera mutación anónima.
  responses:
...
    Unauthorized:
      description: No autenticado / credenciales inválidas
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: No autorizado (rol o propiedad)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Recurso inexistente
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Conflicto de estado (stock, estado de sesión/envío, duplicado, no cancelable)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Error de validación del payload
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Rate limit excedido
      headers:
        Retry-After:
          schema: { type: integer }
          description: Segundos hasta poder reintentar
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Money:
      type: string
      description: Importe decimal con escala fija, serializado como string (BigDecimal).
      examples: ['1500.00']
    Currency:
      type: string
      description: Código de moneda ISO 4217.
      examples: [ARS]
    Problem:
      type: object
      description: Error según RFC 7807.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }
...
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, examples: [Bearer] }
        expiresIn: { type: integer, examples: [900] }
        user: { $ref: '#/components/schemas/UserSummary' }
    UserSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
    Role:
      type: string
      enum: [CUSTOMER, EMPLOYEE, ADMIN]
...
    CartItem:
      type: object
      properties:
        skuId: { type: string, format: uuid }
        sku: { type: [string, 'null'] }
        productId: { type: [string, 'null'], format: uuid }
        productName: { type: [string, 'null'] }
        brand: { type: [string, 'null'] }
        attributes:
          type: object
          additionalProperties: { type: string }
        imageUrl: { type: [string, 'null'], format: uri }
        quantity: { type: integer }
        unitPrice: { $ref: '#/components/schemas/Money' }
        subtotal: { $ref: '#/components/schemas/Money' }
        available: { type: boolean }
    CartItemQuantity:
      type: object
      properties:
        skuId: { type: string, format: uuid }
        quantity: { type: integer }
    Cart:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItem' }
        totals: { $ref: '#/components/schemas/Totals' }
        currency: { $ref: '#/components/schemas/Currency' }
    Totals:
      type: object
      properties:
        items: { $ref: '#/components/schemas/Money' }
        shipping: { $ref: '#/components/schemas/Money' }
        grandTotal: { $ref: '#/components/schemas/Money' }
...